--- src/components/footer.js
@@ -1,13 +1,0 @@
-import * as React from "react";
-
-function Footer() {
-  return (
-    <ul>
-      <li>Email: jetkan.yk@gmail.com</li>
-      <li>GitHub: https://github.com/jetkan-yk</li>
-      <li>LinkedIn: https://www.linkedin.com/in/jet-kan-94916b222/</li>
-    </ul>
-  );
-}
-
-export default Footer;
