+++ src/components/layout/footer.js
@@ -1,0 +1,15 @@
+import * as React from "react";
+
+function Footer() {
+  return (
+    <section id="footer">
+      <ul>
+        <li>Email: jetkan.yk@gmail.com</li>
+        <li>GitHub: https://github.com/jetkan-yk</li>
+        <li>LinkedIn: https://www.linkedin.com/in/jet-kan-94916b222/</li>
+      </ul>
+    </section>
+  );
+}
+
+export default Footer;
