+++ src/components/layout/header.js
@@ -1,0 +1,11 @@
+import * as React from "react";
+import NavBar from "./navbar";
+
+function Header() {
+  return (
+    <section id="header">
+      <NavBar />
+    </section>
+  );
+}
+export default Header;
